refactor: enum -> constant 패키지로 분리
