+++ src/main/java/trend_setter/turtlerun/content/entity/VideoFile.java
@@ -9,6 +9,7 @@ import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.AccessLevel;
+import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 
@@ -22,16 +23,12 @@ public class VideoFile {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Enumerated(EnumType.STRING)
-    private VideoResolution videoResolution;
-
     private String fileName;
     private String filePath;
     private int duration;
 
-    protected VideoFile(VideoResolution videoResolution, String fileName, String filePath,
-        int duration) {
-        this.videoResolution = videoResolution;
+    @Builder
+    protected VideoFile(String fileName, String filePath, int duration) {
         this.fileName = fileName;
         this.filePath = filePath;
         this.duration = duration;
